subscription/artisan_sv/requirements.txt
cryptography
datetime
Pillow
